TGCollector/app/requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
beautifulsoup4==4.12.3
blinker==1.4
certifi==2020.6.20
cffi==1.17.1
chardet==4.0.0
click==8.0.3
cloud-init==24.3.1
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
dbus-python==1.2.18
decorator==5.1.1
distro==1.7.0
distro-info==1.1+ubuntu0.2
elasticsearch==7.13.1
geographiclib==2.0
geopy==2.4.1
geotext==0.4.0
google==3.0.0
gspread==3.1.0
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
jeepney==0.7.1
Jinja2==3.0.3
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.0.1
minio==7.2.9
more-itertools==8.10.0
mysql-connector-python==8.0.18
netifaces==0.11.0
oauth2client==4.1.3
oauthlib==3.2.0
pbr==6.1.0
pexpect==4.8.0
pika==1.3.2
protobuf==4.25.1
ptyprocess==0.7.0
pyaes==1.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycparser==2.22
pycryptodome==3.21.0
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu4
python-debian==0.1.43+ubuntu1.1
python-magic==0.4.24
pytz==2022.1
PyYAML==6.0.1
requests==2.25.1
rsa==4.9
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
sos==4.5.6
soupsieve==2.6
SQLAlchemy==1.3.11
sqlalchemy-migrate==0.13.0
sqlparse==0.5.1
ssh-import-id==5.11
systemd-python==234
Telethon==1.32.0
Tempita==0.5.2
Twisted==22.1.0
typing_extensions==4.12.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
uritools==4.0.1
urllib3==1.26.5
wadllib==1.3.6
xkit==0.0.0
zipp==1.0.0
zope.interface==5.4.0
